requirements.txt
azure-identity==1.15.0
# Flask[async]==2.3.2
openai==1.6.1
azure-search-documents
azure-storage-blob==12.17.0
python-dotenv==1.0.0
azure-cosmos==4.5.0
quart==0.19.4
uvicorn==0.24.0
aiohttp==3.9.2
gunicorn==20.1.0
llama-index
llama-index-llms-azure-openai
llama-index-vector-stores-azureaisearch==0.1.4
llama-index-embeddings-azure-openai